+++ NLayeredApp.Repository/Seeds/ProductFeatureTypeSeed.cs
@@ -11,166 +11,166 @@ public static class ProductFeatureTypeSeed
             // Electronics
             // Galaxy S25
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.GalaxyS25,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Electronics.ColorIds.Silver
+                SeedConstants.ProductIds.Electronics.GalaxyS25,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Electronics.ColorIds.Silver
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.GalaxyS25,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Storage,
-                (int)SeedConstants.FeatureIds.Electronics.StorageIds.Gb256
+                SeedConstants.ProductIds.Electronics.GalaxyS25,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Storage,
+                SeedConstants.FeatureIds.Electronics.StorageIds.Gb256
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.GalaxyS25,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Memory,
-                (int)SeedConstants.FeatureIds.Electronics.MemoryIds.Gb16
+                SeedConstants.ProductIds.Electronics.GalaxyS25,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Memory,
+                SeedConstants.FeatureIds.Electronics.MemoryIds.Gb16
             ),
             
             // Galaxy S24
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.GalaxyS24,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Electronics.ColorIds.White
+                SeedConstants.ProductIds.Electronics.GalaxyS24,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Electronics.ColorIds.White
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.GalaxyS24,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Storage,
-                (int)SeedConstants.FeatureIds.Electronics.StorageIds.Gb128
+                SeedConstants.ProductIds.Electronics.GalaxyS24,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Storage,
+                SeedConstants.FeatureIds.Electronics.StorageIds.Gb128
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.GalaxyS24,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Memory,
-                (int)SeedConstants.FeatureIds.Electronics.MemoryIds.Gb8
+                SeedConstants.ProductIds.Electronics.GalaxyS24,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Memory,
+                SeedConstants.FeatureIds.Electronics.MemoryIds.Gb8
             ),
             
             // Logitech Mouse
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.LogitechMouse,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Electronics.ColorIds.Black
+                SeedConstants.ProductIds.Electronics.LogitechMouse,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Electronics.ColorIds.Black
             ),
             
             // Logitech Keyboard
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.LogitechKeyboard,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Electronics.ColorIds.Black
+                SeedConstants.ProductIds.Electronics.LogitechKeyboard,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Electronics.ColorIds.Black
             ),
             
             // Logitech Headset
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.LogitechHeadset,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Electronics.ColorIds.Black
+                SeedConstants.ProductIds.Electronics.LogitechHeadset,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Electronics.ColorIds.Black
             ),
             
             // Logitech Gamepad
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Electronics.LogitechGamepad,
-                (int)SeedConstants.CategoryIds.Electronics,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Electronics.ColorIds.Silver
+                SeedConstants.ProductIds.Electronics.LogitechGamepad,
+                SeedConstants.CategoryIds.Electronics,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Electronics.ColorIds.Silver
             ),
             
             // Clothing
             // Polo Shirt
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Clothing.PoloShirt,
-                (int)SeedConstants.CategoryIds.Clothing,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Clothing.ColorIds.Red
+                SeedConstants.ProductIds.Clothing.PoloShirt,
+                SeedConstants.CategoryIds.Clothing,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Clothing.ColorIds.Red
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Clothing.PoloShirt,
-                (int)SeedConstants.CategoryIds.Clothing,
-                (int)SeedConstants.FeatureTypeIds.Material,
-                (int)SeedConstants.FeatureIds.Clothing.MaterialIds.Cotton
+                SeedConstants.ProductIds.Clothing.PoloShirt,
+                SeedConstants.CategoryIds.Clothing,
+                SeedConstants.FeatureTypeIds.Material,
+                SeedConstants.FeatureIds.Clothing.MaterialIds.Cotton
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Clothing.PoloShirt,
-                (int)SeedConstants.CategoryIds.Clothing,
-                (int)SeedConstants.FeatureTypeIds.Size,
-                (int)SeedConstants.FeatureIds.Clothing.SizeIds.M
+                SeedConstants.ProductIds.Clothing.PoloShirt,
+                SeedConstants.CategoryIds.Clothing,
+                SeedConstants.FeatureTypeIds.Size,
+                SeedConstants.FeatureIds.Clothing.SizeIds.M
             ),
             
             // Blue Denim Jean
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Clothing.BlueDenimJean,
-                (int)SeedConstants.CategoryIds.Clothing,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Clothing.ColorIds.Blue
+                SeedConstants.ProductIds.Clothing.BlueDenimJean,
+                SeedConstants.CategoryIds.Clothing,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Clothing.ColorIds.Blue
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Clothing.BlueDenimJean,
-                (int)SeedConstants.CategoryIds.Clothing,
-                (int)SeedConstants.FeatureTypeIds.Material,
-                (int)SeedConstants.FeatureIds.Clothing.MaterialIds.Denim
+                SeedConstants.ProductIds.Clothing.BlueDenimJean,
+                SeedConstants.CategoryIds.Clothing,
+                SeedConstants.FeatureTypeIds.Material,
+                SeedConstants.FeatureIds.Clothing.MaterialIds.Denim
             ),
             
             // Running Shoes
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Clothing.RunningShoes,
-                (int)SeedConstants.CategoryIds.Clothing,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.Clothing.ColorIds.Green
+                SeedConstants.ProductIds.Clothing.RunningShoes,
+                SeedConstants.CategoryIds.Clothing,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.Clothing.ColorIds.Green
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.Clothing.RunningShoes,
-                (int)SeedConstants.CategoryIds.Clothing,
-                (int)SeedConstants.FeatureTypeIds.Material,
-                (int)SeedConstants.FeatureIds.Clothing.MaterialIds.Polyester
+                SeedConstants.ProductIds.Clothing.RunningShoes,
+                SeedConstants.CategoryIds.Clothing,
+                SeedConstants.FeatureTypeIds.Material,
+                SeedConstants.FeatureIds.Clothing.MaterialIds.Polyester
             ),
             
             // Home and Living
             // Bookshelf
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.HomeAndLiving.Bookshelf,
-                (int)SeedConstants.CategoryIds.HomeAndLiving,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.HomeAndLiving.ColorIds.Brown
+                SeedConstants.ProductIds.HomeAndLiving.Bookshelf,
+                SeedConstants.CategoryIds.HomeAndLiving,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.HomeAndLiving.ColorIds.Brown
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.HomeAndLiving.Bookshelf,
-                (int)SeedConstants.CategoryIds.HomeAndLiving,
-                (int)SeedConstants.FeatureTypeIds.Material,
-                (int)SeedConstants.FeatureIds.HomeAndLiving.MaterialIds.Wood
+                SeedConstants.ProductIds.HomeAndLiving.Bookshelf,
+                SeedConstants.CategoryIds.HomeAndLiving,
+                SeedConstants.FeatureTypeIds.Material,
+                SeedConstants.FeatureIds.HomeAndLiving.MaterialIds.Wood
             ),
             
             // Dining Table Set
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.HomeAndLiving.DiningTable,
-                (int)SeedConstants.CategoryIds.HomeAndLiving,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.HomeAndLiving.ColorIds.Gray
+                SeedConstants.ProductIds.HomeAndLiving.DiningTable,
+                SeedConstants.CategoryIds.HomeAndLiving,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.HomeAndLiving.ColorIds.Gray
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.HomeAndLiving.DiningTable,
-                (int)SeedConstants.CategoryIds.HomeAndLiving,
-                (int)SeedConstants.FeatureTypeIds.Material,
-                (int)SeedConstants.FeatureIds.HomeAndLiving.MaterialIds.Ceramic
+                SeedConstants.ProductIds.HomeAndLiving.DiningTable,
+                SeedConstants.CategoryIds.HomeAndLiving,
+                SeedConstants.FeatureTypeIds.Material,
+                SeedConstants.FeatureIds.HomeAndLiving.MaterialIds.Ceramic
             ),
             
             // LED Floor Lamp
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.HomeAndLiving.LedFloorLamp,
-                (int)SeedConstants.CategoryIds.HomeAndLiving,
-                (int)SeedConstants.FeatureTypeIds.Color,
-                (int)SeedConstants.FeatureIds.HomeAndLiving.ColorIds.White
+                SeedConstants.ProductIds.HomeAndLiving.LedFloorLamp,
+                SeedConstants.CategoryIds.HomeAndLiving,
+                SeedConstants.FeatureTypeIds.Color,
+                SeedConstants.FeatureIds.HomeAndLiving.ColorIds.White
             ),
             SeedHelpers.CreateProductFeatureType(
-                (int)SeedConstants.ProductIds.HomeAndLiving.LedFloorLamp,
-                (int)SeedConstants.CategoryIds.HomeAndLiving,
-                (int)SeedConstants.FeatureTypeIds.Material,
-                (int)SeedConstants.FeatureIds.HomeAndLiving.MaterialIds.Metal
+                SeedConstants.ProductIds.HomeAndLiving.LedFloorLamp,
+                SeedConstants.CategoryIds.HomeAndLiving,
+                SeedConstants.FeatureTypeIds.Material,
+                SeedConstants.FeatureIds.HomeAndLiving.MaterialIds.Metal
             )
         );
     }
